% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_expanded_neighbours.R
\name{build_expanded_neighbours}
\alias{build_expanded_neighbours}
\title{Build 2D neighbourhood matrix including diagonal neigbours}
\usage{
build_expanded_neighbours(dim1, dim2)
}
\arguments{
\item{dim1}{grid size in x direction}

\item{dim2}{grid size in y direction}
}
\value{
returns a sparse neighbourhood matrix
}
\description{
Build 2D neighbourhood matrix including diagonal neigbours
}
\details{
Builds a 2D neighbourhood matrix as used as D in the genlasso function
}
