% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_lassos.R
\name{build_lassos}
\alias{build_lassos}
\title{Calculate 2D fused-lasso solutionpath for a countmatrix and save as object}
\usage{
build_lassos(counts_matrix, coords_table, name, ncores = 8,
  include_diag = FALSE, output_folder, gamma = 1)
}
\arguments{
\item{counts_matrix}{input matrix of counts with genes in columns and spot identifier in rows}

\item{coords_table}{coordinate table which converts the spot identifier into an x and y coordinate on a grid}

\item{name}{Name of the object to save to.}

\item{ncores}{number of parallel processes to start (default 8)}

\item{include_diag}{wether or not to include diagonal neighbours on the coordinate grid (default FALSE) Warning slow if activated.}

\item{output_folder}{folder in which to save the generated lasso models}

\item{gamma}{ratio of penalization between sparsity and continuity (default 1, high = sparse)}
}
\value{
does not return output but saves to files
}
\description{
Calculate 2D fused-lasso solutionpath for a countmatrix and save as object
}
\details{
This function builds a solution path for each gene (column) in a input matrix and saves it as an .Rdata object
}
