% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fused_lasso_complete_fixed_gamma.R
\name{fused_lasso_complete_fixed_gamma}
\alias{fused_lasso_complete_fixed_gamma}
\title{Building sparse fused lasso models for given genes from their count data}
\usage{
fused_lasso_complete_fixed_gamma(gene_list, counts_matrix, ids_table, name,
  output_folder, include_diag = FALSE, gamma = 1)
}
\arguments{
\item{gene_list}{a set of genen names for which to build the models}

\item{counts_matrix}{a matrix with gene expression levels across genes (columns) and associated spots (rows)}

\item{ids_table}{table  associating each rowname of the counts_matrix with an x and y coordinate}

\item{name}{name of the object to save to}

\item{output_folder}{folder to save object in}

\item{include_diag}{logical value wether to include diagonal spots as neigbours when computing the lasso, Warning slow if active (default FALSE)}

\item{gamma}{ratio of penalization between sparsity and continuity (default 1, high = sparse)}
}
\description{
Building sparse fused lasso models for given genes from their count data
}
\details{
This function will build a sparse fused lasso for each input gene from it's expression vector and the associated spatial coordinates and save the resulting models to an .Rdata object
}
