% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genewise_ggplot.R
\name{genewise_ggplot}
\alias{genewise_ggplot}
\title{Plot Spatial Transcriptomics expression profile for one or more genes}
\usage{
genewise_ggplot(counts_matrix, gene_set, coordinate_matrix,
  only_max = TRUE, z_score = TRUE)
}
\arguments{
\item{counts_matrix}{input matrix of counts with genes in columns and spot identifier in rows}

\item{gene_set}{names of genes to plot, one or more}

\item{coordinate_matrix}{coordinate table which converts the spot identifier into an x and y coordinate on a grid}

\item{only_max}{TURE/FALSE wether to only plot the gene with the highest relative expression for each spot or all gene-expressions on top of each other}

\item{z_score}{TRUE/FALSE wether to normalize expression of genes withint the gene_set by z_score or by dividing by the max expression across al spots}
}
\value{
ggplot2 plot object
}
\description{
Plot Spatial Transcriptomics expression profile for one or more genes
}
\details{
This ggplot2-style function takes a count_matrix and a coordinate_matrix as well as a list of genes whose expression to visualize in 2D using one color per gene
}
