% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_counts.R
\name{load_counts}
\alias{load_counts}
\title{load all count-files from a Spatial Transcriptomics data-set that was processed using simpleST}
\usage{
load_counts(alignments_folder, output_folder)
}
\arguments{
\item{alignments_folder}{path to the parent folder containing one sub-folder per alignment named by their spatial barcodes}

\item{output_folder}{path where to save the R object containing the count_matrix}
}
\value{
does not return output but saves to files
}
\description{
load all count-files from a Spatial Transcriptomics data-set that was processed using simpleST
}
\details{
This function will load the count-files from the 1088 separate sequencing experiments as processed by simpleST and save the resulting count matrix as a .RDS file in the output-folder
}
