% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ids.R
\name{load_ids}
\alias{load_ids}
\title{Building Spatial Transcriptomics coordinate object}
\usage{
load_ids(file_name, output_folder)
}
\arguments{
\item{file_name}{path to the Spatial Transcriptomics id file}

\item{output_folder}{where to save the .RDS object}
}
\value{
does not return output but saves to files
}
\description{
Building Spatial Transcriptomics coordinate object
}
\details{
This function will build a sparse fused lasso for each input gene from it's expression vector and the associated spatial coordinates and save the resulting models to an .Rdata object
}
