% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_lassos_BIC.R
\name{load_lassos_BIC}
\alias{load_lassos_BIC}
\title{load all lasso models optioimized by BIC}
\usage{
load_lassos_BIC(lasso_folder, name)
}
\arguments{
\item{lasso_folder}{path to the directory containing the lasso R objects from the build_lassos function}

\item{name}{to append to the save files}
}
\value{
does not return output but saves to files in working directory
}
\description{
load all lasso models optioimized by BIC
}
\details{
This function loads all models from a folder as built by the build_lassos function optimizes them by BIC and saves a counts object containing a count_matrix, a fits object all the fitted models, a BICs object containing the BIC value for each model and a lls object containing the loglikelihood value of each model.
}
